More re-factoring. Upgraded to Infinispan 5.1, beta in fact. Not operational in a cluster yet, still some transaction issues. Fixed some FindBug warnings.
git-svn-id: http://ikube.googlecode.com/svn/trunk@700 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/listener/MonitoringListener.java
@@ -32,7 +32,7 @@ public class MonitoringListener implements IListener {
 		for (Execution execution : executions.values()) {
 			long duration = TimeUnit.NANOSECONDS.toSeconds(execution.getDuration());
 			if (duration > 0) {
-				execution.setExecutionsPerSecond((double) (execution.getInvocations() / duration));
+				execution.setExecutionsPerSecond(((double) execution.getInvocations() / (double) duration));
 			}
 		}
 	}
